@@ -1,37 +1,32 @@
 {
-  "_args": [
-    [
-      "qs@6.5.2",
-      "C:\\Users\\jeff\\Desktop\\Artistsacumilated\\Project3"
-    ]
-  ],
-  "_from": "qs@6.5.2",
-  "_id": "qs@6.5.2",
+  "_from": "qs@6.7.0",
+  "_id": "qs@6.7.0",
   "_inBundle": false,
-  "_integrity": "sha512-N5ZAX4/LxJmF+7wN74pUD6qAh9/wnvdQcjq9TZjevvXzSUo7bfmw91saqMjzGS2xq91/odN2dW/WOl7qQHNDGA==",
+  "_integrity": "sha512-VCdBRNFTX1fyE7Nb6FYoURo/SPe62QCaAyzJvUjwRaIsc+NePBEniHlvxFmmX56+HZphIGtV0XeCirBtpDrTyQ==",
   "_location": "/qs",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "qs@6.5.2",
+    "raw": "qs@6.7.0",
     "name": "qs",
     "escapedName": "qs",
-    "rawSpec": "6.5.2",
+    "rawSpec": "6.7.0",
     "saveSpec": null,
-    "fetchSpec": "6.5.2"
+    "fetchSpec": "6.7.0"
   },
   "_requiredBy": [
     "/body-parser",
-    "/express",
-    "/request"
+    "/express"
   ],
-  "_resolved": "https://registry.npmjs.org/qs/-/qs-6.5.2.tgz",
-  "_spec": "6.5.2",
-  "_where": "C:\\Users\\jeff\\Desktop\\Artistsacumilated\\Project3",
+  "_resolved": "https://registry.npmjs.org/qs/-/qs-6.7.0.tgz",
+  "_shasum": "41dc1a015e3d581f1621776be31afb2876a9b1bc",
+  "_spec": "qs@6.7.0",
+  "_where": "C:\\Users\\jeff\\Desktop\\Artistsacumilated\\Project3\\node_modules\\express",
   "bugs": {
     "url": "https://github.com/ljharb/qs/issues"
   },
+  "bundleDependencies": false,
   "contributors": [
     {
       "name": "Jordan Harband",
@@ -40,20 +35,23 @@
     }
   ],
   "dependencies": {},
+  "deprecated": false,
   "description": "A querystring parser that supports nesting and arrays, with a depth limit",
   "devDependencies": {
-    "@ljharb/eslint-config": "^12.2.1",
-    "browserify": "^16.2.0",
-    "covert": "^1.1.0",
+    "@ljharb/eslint-config": "^13.1.1",
+    "browserify": "^16.2.3",
+    "covert": "^1.1.1",
     "editorconfig-tools": "^0.1.1",
-    "eslint": "^4.19.1",
+    "eslint": "^5.15.3",
     "evalmd": "^0.0.17",
-    "iconv-lite": "^0.4.21",
+    "for-each": "^0.3.3",
+    "iconv-lite": "^0.4.24",
     "mkdirp": "^0.5.1",
+    "object-inspect": "^1.6.0",
     "qs-iconv": "^1.0.4",
-    "safe-publish-latest": "^1.1.1",
+    "safe-publish-latest": "^1.1.2",
     "safer-buffer": "^2.1.2",
-    "tape": "^4.9.0"
+    "tape": "^4.10.1"
   },
   "engines": {
     "node": ">=0.6"
@@ -61,7 +59,11 @@
   "homepage": "https://github.com/ljharb/qs",
   "keywords": [
     "querystring",
-    "qs"
+    "qs",
+    "query",
+    "url",
+    "parse",
+    "stringify"
   ],
   "license": "BSD-3-Clause",
   "main": "lib/index.js",
@@ -74,12 +76,12 @@
     "coverage": "covert test",
     "dist": "mkdirp dist && browserify --standalone Qs lib/index.js > dist/qs.js",
     "lint": "eslint lib/*.js test/*.js",
-    "prelint": "editorconfig-tools check * lib/* test/*",
+    "postlint": "editorconfig-tools check * lib/* test/*",
     "prepublish": "safe-publish-latest && npm run dist",
     "pretest": "npm run --silent readme && npm run --silent lint",
     "readme": "evalmd README.md",
     "test": "npm run --silent coverage",
     "tests-only": "node test"
   },
-  "version": "6.5.2"
+  "version": "6.7.0"
 }
