@@ -1,33 +1,28 @@
 {
-  "_args": [
-    [
-      "send@0.16.2",
-      "C:\\Users\\jeff\\Desktop\\Artistsacumilated\\Project3"
-    ]
-  ],
-  "_from": "send@0.16.2",
-  "_id": "send@0.16.2",
+  "_from": "send@0.17.1",
+  "_id": "send@0.17.1",
   "_inBundle": false,
-  "_integrity": "sha512-E64YFPUssFHEFBvpbbjr44NCLtI1AohxQ8ZSiJjQLskAdKuriYEP6VyGEsRDH8ScozGpkaX1BGvhanqCwkcEZw==",
+  "_integrity": "sha512-BsVKsiGcQMFwT8UxypobUKyv7irCNRHk1T0G680vk88yf6LBByGcZJOTJCrTP2xVN6yI+XjPJcNuE3V4fT9sAg==",
   "_location": "/send",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "send@0.16.2",
+    "raw": "send@0.17.1",
     "name": "send",
     "escapedName": "send",
-    "rawSpec": "0.16.2",
+    "rawSpec": "0.17.1",
     "saveSpec": null,
-    "fetchSpec": "0.16.2"
+    "fetchSpec": "0.17.1"
   },
   "_requiredBy": [
     "/express",
     "/serve-static"
   ],
-  "_resolved": "https://registry.npmjs.org/send/-/send-0.16.2.tgz",
-  "_spec": "0.16.2",
-  "_where": "C:\\Users\\jeff\\Desktop\\Artistsacumilated\\Project3",
+  "_resolved": "https://registry.npmjs.org/send/-/send-0.17.1.tgz",
+  "_shasum": "c1d8b059f7900f7466dd4938bdc44e11ddb376c8",
+  "_spec": "send@0.17.1",
+  "_where": "C:\\Users\\jeff\\Desktop\\Artistsacumilated\\Project3\\node_modules\\express",
   "author": {
     "name": "TJ Holowaychuk",
     "email": "tj@vision-media.ca"
@@ -35,6 +30,7 @@
   "bugs": {
     "url": "https://github.com/pillarjs/send/issues"
   },
+  "bundleDependencies": false,
   "contributors": [
     {
       "name": "Douglas Christopher Wilson",
@@ -57,26 +53,27 @@
     "escape-html": "~1.0.3",
     "etag": "~1.8.1",
     "fresh": "0.5.2",
-    "http-errors": "~1.6.2",
-    "mime": "1.4.1",
-    "ms": "2.0.0",
+    "http-errors": "~1.7.2",
+    "mime": "1.6.0",
+    "ms": "2.1.1",
     "on-finished": "~2.3.0",
-    "range-parser": "~1.2.0",
-    "statuses": "~1.4.0"
+    "range-parser": "~1.2.1",
+    "statuses": "~1.5.0"
   },
+  "deprecated": false,
   "description": "Better streaming static file server with Range and conditional-GET support",
   "devDependencies": {
     "after": "0.8.2",
-    "eslint": "3.19.0",
-    "eslint-config-standard": "10.2.1",
-    "eslint-plugin-import": "2.8.0",
-    "eslint-plugin-markdown": "1.0.0-beta.6",
-    "eslint-plugin-node": "5.2.1",
-    "eslint-plugin-promise": "3.6.0",
-    "eslint-plugin-standard": "3.0.1",
+    "eslint": "5.16.0",
+    "eslint-config-standard": "12.0.0",
+    "eslint-plugin-import": "2.17.2",
+    "eslint-plugin-markdown": "1.0.0",
+    "eslint-plugin-node": "8.0.1",
+    "eslint-plugin-promise": "4.1.1",
+    "eslint-plugin-standard": "4.0.0",
     "istanbul": "0.4.5",
-    "mocha": "2.5.3",
-    "supertest": "1.1.0"
+    "mocha": "6.1.4",
+    "supertest": "4.0.2"
   },
   "engines": {
     "node": ">= 0.8.0"
@@ -105,5 +102,5 @@
     "test-ci": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --check-leaks --reporter spec",
     "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --check-leaks --reporter dot"
   },
-  "version": "0.16.2"
+  "version": "0.17.1"
 }
