@@ -1,32 +1,27 @@
 {
-  "_args": [
-    [
-      "serve-static@1.13.2",
-      "C:\\Users\\jeff\\Desktop\\Artistsacumilated\\Project3"
-    ]
-  ],
-  "_from": "serve-static@1.13.2",
-  "_id": "serve-static@1.13.2",
+  "_from": "serve-static@1.14.1",
+  "_id": "serve-static@1.14.1",
   "_inBundle": false,
-  "_integrity": "sha512-p/tdJrO4U387R9oMjb1oj7qSMaMfmOyd4j9hOFoxZe2baQszgHcSWjuya/CiT5kgZZKRudHNOA0pYXOl8rQ5nw==",
+  "_integrity": "sha512-JMrvUwE54emCYWlTI+hGrGv5I8dEwmco/00EvkzIIsR7MqrHonbD9pO2MOfFnpFntl7ecpZs+3mW+XbQZu9QCg==",
   "_location": "/serve-static",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "serve-static@1.13.2",
+    "raw": "serve-static@1.14.1",
     "name": "serve-static",
     "escapedName": "serve-static",
-    "rawSpec": "1.13.2",
+    "rawSpec": "1.14.1",
     "saveSpec": null,
-    "fetchSpec": "1.13.2"
+    "fetchSpec": "1.14.1"
   },
   "_requiredBy": [
     "/express"
   ],
-  "_resolved": "https://registry.npmjs.org/serve-static/-/serve-static-1.13.2.tgz",
-  "_spec": "1.13.2",
-  "_where": "C:\\Users\\jeff\\Desktop\\Artistsacumilated\\Project3",
+  "_resolved": "https://registry.npmjs.org/serve-static/-/serve-static-1.14.1.tgz",
+  "_shasum": "666e636dc4f010f7ef29970a88a674320898b2f9",
+  "_spec": "serve-static@1.14.1",
+  "_where": "C:\\Users\\jeff\\Desktop\\Artistsacumilated\\Project3\\node_modules\\express",
   "author": {
     "name": "Douglas Christopher Wilson",
     "email": "doug@somethingdoug.com"
@@ -34,24 +29,27 @@
   "bugs": {
     "url": "https://github.com/expressjs/serve-static/issues"
   },
+  "bundleDependencies": false,
   "dependencies": {
     "encodeurl": "~1.0.2",
     "escape-html": "~1.0.3",
-    "parseurl": "~1.3.2",
-    "send": "0.16.2"
+    "parseurl": "~1.3.3",
+    "send": "0.17.1"
   },
+  "deprecated": false,
   "description": "Serve static files",
   "devDependencies": {
-    "eslint": "3.19.0",
-    "eslint-config-standard": "10.2.1",
-    "eslint-plugin-import": "2.8.0",
-    "eslint-plugin-markdown": "1.0.0-beta.6",
-    "eslint-plugin-node": "5.2.1",
-    "eslint-plugin-promise": "3.6.0",
-    "eslint-plugin-standard": "3.0.1",
+    "eslint": "5.16.0",
+    "eslint-config-standard": "12.0.0",
+    "eslint-plugin-import": "2.17.2",
+    "eslint-plugin-markdown": "1.0.0",
+    "eslint-plugin-node": "8.0.1",
+    "eslint-plugin-promise": "4.1.1",
+    "eslint-plugin-standard": "4.0.0",
     "istanbul": "0.4.5",
-    "mocha": "2.5.3",
-    "supertest": "1.1.0"
+    "mocha": "6.1.4",
+    "safe-buffer": "5.1.2",
+    "supertest": "4.0.2"
   },
   "engines": {
     "node": ">= 0.8.0"
@@ -72,7 +70,8 @@
     "lint": "eslint --plugin markdown --ext js,md .",
     "test": "mocha --reporter spec --bail --check-leaks test/",
     "test-ci": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter spec --check-leaks test/",
-    "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot --check-leaks test/"
+    "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot --check-leaks test/",
+    "version": "node scripts/version-history.js && git add HISTORY.md"
   },
-  "version": "1.13.2"
+  "version": "1.14.1"
 }
