@@ -1,38 +1,35 @@
 {
-  "_args": [
-    [
-      "statuses@1.4.0",
-      "C:\\Users\\jeff\\Desktop\\Artistsacumilated\\Project3"
-    ]
-  ],
-  "_from": "statuses@1.4.0",
-  "_id": "statuses@1.4.0",
+  "_from": "statuses@~1.5.0",
+  "_id": "statuses@1.5.0",
   "_inBundle": false,
-  "_integrity": "sha512-zhSCtt8v2NDrRlPQpCNtw/heZLtfUDqxBM1udqikb/Hbk52LK4nQSwr10u77iopCW5LsyHpuXS0GnEc48mLeew==",
+  "_integrity": "sha1-Fhx9rBd2Wf2YEfQ3cfqZOBR4Yow=",
   "_location": "/statuses",
   "_phantomChildren": {},
   "_requested": {
-    "type": "version",
+    "type": "range",
     "registry": true,
-    "raw": "statuses@1.4.0",
+    "raw": "statuses@~1.5.0",
     "name": "statuses",
     "escapedName": "statuses",
-    "rawSpec": "1.4.0",
+    "rawSpec": "~1.5.0",
     "saveSpec": null,
-    "fetchSpec": "1.4.0"
+    "fetchSpec": "~1.5.0"
   },
   "_requiredBy": [
     "/express",
     "/finalhandler",
     "/http-errors",
-    "/send"
+    "/send",
+    "/serve-index/http-errors"
   ],
-  "_resolved": "https://registry.npmjs.org/statuses/-/statuses-1.4.0.tgz",
-  "_spec": "1.4.0",
-  "_where": "C:\\Users\\jeff\\Desktop\\Artistsacumilated\\Project3",
+  "_resolved": "https://registry.npmjs.org/statuses/-/statuses-1.5.0.tgz",
+  "_shasum": "161c7dac177659fd9811f43771fa99381478628c",
+  "_spec": "statuses@~1.5.0",
+  "_where": "C:\\Users\\jeff\\Desktop\\Artistsacumilated\\Project3\\node_modules\\express",
   "bugs": {
     "url": "https://github.com/jshttp/statuses/issues"
   },
+  "bundleDependencies": false,
   "contributors": [
     {
       "name": "Douglas Christopher Wilson",
@@ -44,15 +41,16 @@
       "url": "http://jongleberry.com"
     }
   ],
+  "deprecated": false,
   "description": "HTTP status utility",
   "devDependencies": {
     "csv-parse": "1.2.4",
-    "eslint": "3.19.0",
-    "eslint-config-standard": "10.2.1",
-    "eslint-plugin-import": "2.8.0",
+    "eslint": "4.19.1",
+    "eslint-config-standard": "11.0.0",
+    "eslint-plugin-import": "2.9.0",
     "eslint-plugin-markdown": "1.0.0-beta.6",
-    "eslint-plugin-node": "5.2.0",
-    "eslint-plugin-promise": "3.6.0",
+    "eslint-plugin-node": "6.0.1",
+    "eslint-plugin-promise": "3.7.0",
     "eslint-plugin-standard": "3.0.1",
     "istanbul": "0.4.5",
     "mocha": "1.21.5",
@@ -89,5 +87,5 @@
     "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot --check-leaks test/",
     "update": "npm run fetch && npm run build"
   },
-  "version": "1.4.0"
+  "version": "1.5.0"
 }
